requirements.txt
line-bot-sdk=3.7
flask=3.0.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
Werkzeug==3.1.3

